+++ src/ch.usi.inf.nodeprof.test/src/ch/usi/inf/nodeprof/test/examples/BranchCoverage.java
@@ -79,10 +79,10 @@ public class BranchCoverage extends TestableNodeProfAnalysis {
     public void printResult() {
         int idx = 0;
         Logger.info("BranchCoverage analysis finishes.");
-        for (Entry<Long, Report> entry : trueDB.getRecords().entrySet()) {
+        for (Entry<Integer, Report> entry : trueDB.getRecords().entrySet()) {
             Logger.info("BranchCoverage[" + idx++ + "]: True " + entry.getValue().report());
         }
-        for (Entry<Long, Report> entry : falseDB.getRecords().entrySet()) {
+        for (Entry<Integer, Report> entry : falseDB.getRecords().entrySet()) {
             Logger.info("BranchCoverage[" + idx++ + "]: False " + entry.getValue().report());
         }
     }
